Introduce an improved AES256 cipher version
Now we don't add spurious padding for AES-CTR (which is used as a stream
cipher, and doesn't need padding) and we use base64 to encode the output
message instead of blowing up its size by hexlify()ing everything twice.
We also don't unnecessarily use PBKDF2 to generate the initial value for
the encryption counter.

We always use the new format for output, but maintain read-compatibility
with a few lines of 1.1-specific decoding in decrypt(). (Note that we do
not force v1.1 files to be rewritten by ansible-vault edit unless their
contents have actually changed, in which case the new cipher version is
used automatically).

Closes #10634 from graingert (which removed the padding, but added yet
another cipher class for backwards compatibility)

Fixes #12121 from mgedmin (which reported the double-hexlify() problem)