Issue doas sh -c 'echo BECOME-SUCCESS-…; /usr/bin/python …'
doas doesn't remember successful authentications the way sudo does, so
the "doas a && doas b" command we were issuing earlier would not work,
because doas would ask for a password twice, and the ssh plugin could
not deal with that.

Also don't bother to set ANSIBLE=true, since nothing seems to check.

Tested on OpenBSD 5.8.